Move WHO form tests to a new smoke-test module
